giveaway and events persistence

+++ src/main/groovy/pl/jug/torun/xenia/rest/dto/GiveAwayRequest.groovy
@@ -1,9 +1,12 @@
 package pl.jug.torun.xenia.rest.dto
 
+import pl.jug.torun.xenia.model.GiveAway
+
 /**
  * Created by mephi_000 on 06.09.14.
  */
 class GiveAwayRequest {
-    int prizeId
+    long prizeId
     int amount
+   
 }
